+++ java/src/org/chromium/media/MediaPlayerListener.java
@@ -1,12 +1,10 @@
-// Copyright (c) 2012 The Chromium Authors. All rights reserved.
+// Copyright 2012 The Chromium Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
 package org.chromium.media;
 
-import android.Manifest.permission;
 import android.content.Context;
-import android.content.pm.PackageManager;
 import android.media.AudioManager;
 import android.media.MediaPlayer;
 
@@ -35,10 +33,10 @@ class MediaPlayerListener implements MediaPlayer.OnPreparedListener,
     public static final int MEDIA_ERROR_TIMED_OUT = -110;
 
     // Used to determine the class instance to dispatch the native call to.
-    private int mNativeMediaPlayerListener = 0;
+    private long mNativeMediaPlayerListener = 0;
     private final Context mContext;
 
-    private MediaPlayerListener(int nativeMediaPlayerListener, Context context) {
+    private MediaPlayerListener(long nativeMediaPlayerListener, Context context) {
         mNativeMediaPlayerListener = nativeMediaPlayerListener;
         mContext = context;
     }
@@ -148,22 +146,22 @@ class MediaPlayerListener implements MediaPlayer.OnPreparedListener,
      * See media/base/android/media_player_listener.cc for all the following functions.
      */
     private native void nativeOnMediaError(
-            int nativeMediaPlayerListener,
+            long nativeMediaPlayerListener,
             int errorType);
 
     private native void nativeOnVideoSizeChanged(
-            int nativeMediaPlayerListener,
+            long nativeMediaPlayerListener,
             int width, int height);
 
     private native void nativeOnBufferingUpdate(
-            int nativeMediaPlayerListener,
+            long nativeMediaPlayerListener,
             int percent);
 
-    private native void nativeOnMediaPrepared(int nativeMediaPlayerListener);
+    private native void nativeOnMediaPrepared(long nativeMediaPlayerListener);
 
-    private native void nativeOnPlaybackComplete(int nativeMediaPlayerListener);
+    private native void nativeOnPlaybackComplete(long nativeMediaPlayerListener);
 
-    private native void nativeOnSeekComplete(int nativeMediaPlayerListener);
+    private native void nativeOnSeekComplete(long nativeMediaPlayerListener);
 
-    private native void nativeOnMediaInterrupted(int nativeMediaPlayerListener);
+    private native void nativeOnMediaInterrupted(long nativeMediaPlayerListener);
 }
